# dist/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:8000/api/v20230821",
      "description": "Local Server"
    },
    {
      "url": "http://localhost:3030/api/v20230821",
      "description": "Local Server no docker"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "oduck-server",
    "description": ""
  },
  "paths": {
    "/animation": {
      "post": {
        "tags": [
          "admin/animation"
        ],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimationReqDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimationItemResDto"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "@tag admin/animation",
        "x-nestia-namespace": "animation.store",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "admin/animation",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "POST"
      },
      "get": {
        "tags": [
          "animation"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/AnimationListDto"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimationItemResDto"
                  }
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "@tag animation",
        "x-nestia-namespace": "animation.getList",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "animation",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "GET"
      }
    },
    "/animation/{id}": {
      "put": {
        "tags": [
          "admin/animation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Partial_lt_AnimationBaseDto_gt_"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResult_lt__blt__space_id:_space_number;_space_name:_space_string;_space_plot:_space_string;_space_broadcastType:_space_BroadcastType;_space_episodeNumber:_space_number;_space_imageUrl:_space_string;_space_rating:_space_Rating;_space_primaryKeyword:_space_string;_space_status:_space_Status;_space_..._space_5_space_more_space_...;_space_deletedAt:_space_Date_space__or__space_null;_space__bgt__comma__space_unknown_gt__space__and__space__blt__bgt_"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "@tag admin/animation",
        "x-nestia-namespace": "animation.update",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "admin/animation",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "PUT"
      },
      "delete": {
        "tags": [
          "admin/animation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResult_lt__blt__space_id:_space_number;_space_name:_space_string;_space_plot:_space_string;_space_broadcastType:_space_BroadcastType;_space_episodeNumber:_space_number;_space_imageUrl:_space_string;_space_rating:_space_Rating;_space_primaryKeyword:_space_string;_space_status:_space_Status;_space_..._space_5_space_more_space_...;_space_deletedAt:_space_Date_space__or__space_null;_space__bgt__comma__space_unknown_gt__space__and__space__blt__bgt_"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "@tag admin/animation",
        "x-nestia-namespace": "animation.destroy",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "admin/animation",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "DELETE"
      },
      "get": {
        "tags": [
          "animation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimationItemResDto"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "@tag animation",
        "x-nestia-namespace": "animation.show",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "animation",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "GET"
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnimationReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "@default 바키",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "바키",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "default": "바키"
          },
          "plot": {
            "description": "@default 아부지 싸움 잘하시네요?",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "아부지 싸움 잘하시네요?",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "default": "아부지 싸움 잘하시네요?"
          },
          "broadcastType": {
            "$ref": "#/components/schemas/BroadcastType"
          },
          "episodeNumber": {
            "description": "@default 1",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "1",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number",
            "default": 1
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "primaryKeyword": {
            "description": "@default 싸움",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "싸움",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "default": "싸움"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "isReleased": {
            "description": "@default true",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "true",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "default": true,
            "type": "boolean"
          },
          "imageUrl": {
            "description": "@default https://placehold.co/600x400",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "https://placehold.co/600x400",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "default": "https://placehold.co/600x400"
          },
          "studioNames": {
            "description": "@default 가이낙스",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "가이낙스",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "description": "@default 가이낙스",
              "x-typia-jsDocTags": [
                {
                  "name": "default",
                  "text": [
                    {
                      "text": "가이낙스",
                      "kind": "text"
                    }
                  ]
                }
              ],
              "x-typia-required": true,
              "x-typia-optional": false,
              "type": "string",
              "default": "가이낙스"
            }
          }
        },
        "nullable": false,
        "required": [
          "name",
          "plot",
          "broadcastType",
          "episodeNumber",
          "rating",
          "primaryKeyword",
          "status",
          "isReleased",
          "imageUrl",
          "studioNames"
        ],
        "x-typia-jsDocTags": []
      },
      "BroadcastType": {
        "type": "string",
        "enum": [
          "TVA",
          "OVA",
          "ONA",
          "MOV"
        ]
      },
      "Rating": {
        "type": "string",
        "enum": [
          "ADULT",
          "FIFTEEN",
          "TWELVE",
          "ALL"
        ]
      },
      "Status": {
        "type": "string",
        "enum": [
          "FINISHED",
          "ONGOING",
          "UPCOMING",
          "UNKNOWN"
        ]
      },
      "AnimationItemResDto": {
        "type": "object",
        "properties": {
          "studios": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetResult_lt__blt__space_id:_space_number;_space_name:_space_string;_space_createdAt:_space_Date;_space_updatedAt:_space_Date;_space__bgt__comma__space_unknown_gt__space__and__space__blt__bgt_"
            }
          },
          "name": {
            "description": "@default 바키",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "바키",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "바키"
          },
          "plot": {
            "description": "@default 아부지 싸움 잘하시네요?",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "아부지 싸움 잘하시네요?",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "아부지 싸움 잘하시네요?"
          },
          "broadcastType": {
            "description": "@default OVA",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "OVA",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "TVA",
              "OVA",
              "ONA",
              "MOV"
            ],
            "default": "OVA"
          },
          "episodeNumber": {
            "description": "@default 1",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "1",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "number",
            "default": 1
          },
          "rating": {
            "description": "@default ADULT",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "ADULT",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "ADULT",
              "FIFTEEN",
              "TWELVE",
              "ALL"
            ],
            "default": "ADULT"
          },
          "primaryKeyword": {
            "description": "@default 싸움",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "싸움",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "싸움"
          },
          "status": {
            "description": "@default ONGOING",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "ONGOING",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "FINISHED",
              "ONGOING",
              "UPCOMING",
              "UNKNOWN"
            ],
            "default": "ONGOING"
          },
          "isReleased": {
            "description": "@default true",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "true",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "default": true,
            "type": "boolean"
          },
          "imageUrl": {
            "description": "@default https://placehold.co/600x400",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "https://placehold.co/600x400",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "https://placehold.co/600x400"
          }
        },
        "nullable": false,
        "required": [
          "studios"
        ],
        "x-typia-jsDocTags": []
      },
      "GetResult_lt__blt__space_id:_space_number;_space_name:_space_string;_space_createdAt:_space_Date;_space_updatedAt:_space_Date;_space__bgt__comma__space_unknown_gt__space__and__space__blt__bgt_": {
        "type": "object",
        "properties": {
          "id": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "name": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "createdAt": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "date-time"
          }
        },
        "nullable": false,
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "x-typia-jsDocTags": []
      },
      "Partial_lt_AnimationBaseDto_gt_": {
        "type": "object",
        "properties": {
          "name": {
            "description": "@default 바키",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "바키",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "바키"
          },
          "plot": {
            "description": "@default 아부지 싸움 잘하시네요?",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "아부지 싸움 잘하시네요?",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "아부지 싸움 잘하시네요?"
          },
          "broadcastType": {
            "description": "@default OVA",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "OVA",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "TVA",
              "OVA",
              "ONA",
              "MOV"
            ],
            "default": "OVA"
          },
          "episodeNumber": {
            "description": "@default 1",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "1",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "number",
            "default": 1
          },
          "rating": {
            "description": "@default ADULT",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "ADULT",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "ADULT",
              "FIFTEEN",
              "TWELVE",
              "ALL"
            ],
            "default": "ADULT"
          },
          "primaryKeyword": {
            "description": "@default 싸움",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "싸움",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "싸움"
          },
          "status": {
            "description": "@default ONGOING",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "ONGOING",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "enum": [
              "FINISHED",
              "ONGOING",
              "UPCOMING",
              "UNKNOWN"
            ],
            "default": "ONGOING"
          },
          "isReleased": {
            "description": "@default true",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "true",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "default": true,
            "type": "boolean"
          },
          "imageUrl": {
            "description": "@default https://placehold.co/600x400",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "https://placehold.co/600x400",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": false,
            "type": "string",
            "default": "https://placehold.co/600x400"
          }
        },
        "nullable": false,
        "x-typia-jsDocTags": []
      },
      "GetResult_lt__blt__space_id:_space_number;_space_name:_space_string;_space_plot:_space_string;_space_broadcastType:_space_BroadcastType;_space_episodeNumber:_space_number;_space_imageUrl:_space_string;_space_rating:_space_Rating;_space_primaryKeyword:_space_string;_space_status:_space_Status;_space_..._space_5_space_more_space_...;_space_deletedAt:_space_Date_space__or__space_null;_space__bgt__comma__space_unknown_gt__space__and__space__blt__bgt_": {
        "type": "object",
        "properties": {
          "id": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "name": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "plot": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "broadcastType": {
            "$ref": "#/components/schemas/BroadcastType"
          },
          "episodeNumber": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "imageUrl": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "primaryKeyword": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "isReleased": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "boolean"
          },
          "viewCount": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "reviewCount": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "createdAt": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "date-time"
          }
        },
        "nullable": false,
        "required": [
          "id",
          "name",
          "plot",
          "broadcastType",
          "episodeNumber",
          "imageUrl",
          "rating",
          "primaryKeyword",
          "status",
          "isReleased",
          "viewCount",
          "reviewCount",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ],
        "x-typia-jsDocTags": []
      },
      "AnimationListDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "description": "@default 20",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "20",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": true,
            "type": "number",
            "default": 20
          },
          "lastId": {
            "description": "@description 직전 페이지 마지막 ID / empty '' 면 처음부터",
            "x-typia-jsDocTags": [
              {
                "name": "description",
                "text": [
                  {
                    "text": "직전 페이지 마지막 ID / empty '' 면 처음부터",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": true,
            "type": "number"
          },
          "sortBy": {
            "description": "@default createdAt",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "createdAt",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": true,
            "type": "string",
            "default": "createdAt"
          },
          "sortOrder": {
            "description": "@default desc\n@enum Sort",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "desc",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "enum"
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": true,
            "type": "string",
            "default": "desc"
          },
          "search": {
            "description": "@default 바키",
            "x-typia-jsDocTags": [
              {
                "name": "default",
                "text": [
                  {
                    "text": "바키",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "x-typia-optional": true,
            "type": "string",
            "default": "바키"
          }
        },
        "nullable": false,
        "description": "@default 바키",
        "x-typia-jsDocTags": []
      },
...
    },
    "securitySchemes": {
      "apiCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oDuckio.sid"
      }
    }
  },
  "security": [
    {
      "apiCookie": []
    }
  ]
}